Extract debug and disassembly to debug.go
