+++ cpu.go
@@ -315,103 +315,3 @@ func (c *CPU) GetCurrentInstruction() *Instruction {
 func (c *CPU) Opcode() uint8 {
 	return c.opcode
 }
-
-// Disassemble - Disassemble instructions in memory range
-func (c *CPU) Disassemble(startAddr, endAddr uint16) map[uint16]string {
-	disassembly := make(map[uint16]string)
-	addr := startAddr
-
-	for addr <= endAddr && addr >= startAddr { // Check for wrap around too
-		lineAddr := addr
-		opcode := c.read(addr)
-		addr++ // Consume opcode byte
-		instr := c.lookup[opcode]
-		operandStr := ""
-
-		switch instr.AddrModeType { // Use enum comparison
-		case AddrModeIMP:
-			// No operand bytes
-		case AddrModeIMM:
-			val := c.read(addr)
-			addr++
-			operandStr = fmt.Sprintf("#$%02X", val)
-		case AddrModeZP0:
-			val := c.read(addr)
-			addr++
-			operandStr = fmt.Sprintf("$%02X", val)
-		case AddrModeZPX:
-			val := c.read(addr)
-			addr++
-			operandStr = fmt.Sprintf("$%02X,X", val)
-		case AddrModeZPY:
-			val := c.read(addr)
-			addr++
-			operandStr = fmt.Sprintf("$%02X,Y", val)
-		case AddrModeREL:
-			val := c.read(addr)
-			addr++            // Consume operand byte
-			relTarget := addr // Relative jump target is relative to the *next* instruction
-			if val&0x80 > 0 {
-				relTarget += uint16(int16(val)) // Handle negative offset
-			} else {
-				relTarget += uint16(val) // Positive offset
-			}
-			operandStr = fmt.Sprintf("$%04X", relTarget)
-		case AddrModeABS:
-			lo := uint16(c.read(addr))
-			addr++
-			hi := uint16(c.read(addr))
-			addr++
-			operandStr = fmt.Sprintf("$%04X", (hi<<8)|lo)
-		case AddrModeABX:
-			lo := uint16(c.read(addr))
-			addr++
-			hi := uint16(c.read(addr))
-			addr++
-			operandStr = fmt.Sprintf("$%04X,X", (hi<<8)|lo)
-		case AddrModeABY:
-			lo := uint16(c.read(addr))
-			addr++
-			hi := uint16(c.read(addr))
-			addr++
-			operandStr = fmt.Sprintf("$%04X,Y", (hi<<8)|lo)
-		case AddrModeIND:
-			lo := uint16(c.read(addr))
-			addr++
-			hi := uint16(c.read(addr))
-			addr++
-			operandStr = fmt.Sprintf("($%04X)", (hi<<8)|lo)
-		case AddrModeIZX:
-			val := c.read(addr)
-			addr++
-			operandStr = fmt.Sprintf("($%02X,X)", val)
-		case AddrModeIZY:
-			val := c.read(addr)
-			addr++
-			operandStr = fmt.Sprintf("($%02X),Y", val)
-		default:
-			operandStr = "???"
-		}
-
-		// Add safety checks in case addr went beyond bounds during operand read
-		if addr > endAddr+4 {
-			disassembly[lineAddr] = fmt.Sprintf("%s ??? ; Address bounds exceeded", instr.Name)
-			break
-		}
-
-		disassembly[lineAddr] = fmt.Sprintf("%s %s", instr.Name, operandStr)
-
-		// Safety break for very large ranges or infinite loops from bad data
-		if addr < lineAddr && lineAddr != 0 {
-			break
-		} // Address wrapped around negatively
-
-	}
-	return disassembly
-}
-
-// LookupTable exposes the instruction lookup table (use with caution).
-// Primarily intended for tools like disassemblers or UI displays.
-func (c *CPU) LookupTable() [256]Instruction {
-	return c.lookup
-}
